Change int64_t  to double

+++ packages/audioplayers_windows/windows/audioplayers_windows_plugin.cpp
@@ -155,7 +155,7 @@ void AudioplayersWindowsPlugin::HandleMethodCall(
     player->SeekTo(0);
     result->Success(EncodableValue(1));
   } else if (method_call.method_name().compare("seek") == 0) {
-    auto position = GetArgument<int64_t>("position", args, (int64_t)(player->GetPosition()));
+    auto position = GetArgument<double>("position", args, (double)(player->GetPosition()));
     player->SeekTo(static_cast<int64_t>(position));
     result->Success(EncodableValue(1));
   } else if (method_call.method_name().compare("setSourceUrl") == 0) {
